Implemented Ruby 3.2 support; changed tests to include less 2.7.7 but cover 3.2.2
