+++ .github/workflows/gem-test-and-release.yml
@@ -105,7 +105,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ['alpine:3.16']
-        package_ruby_ver: [ '3.0.6' ]
+        package_ruby_ver: [ '3.1.4' ]
         env:
           - CC: gcc
             CXX: g++
@@ -182,7 +182,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [  macos-11.0  ]
-        package_ruby_ver: [ '3.0.6' ]
+        package_ruby_ver: [ '3.1.4' ]
         env:
           - CC: clang
             CXX: clang++
@@ -240,7 +240,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ ubuntu-20.04 ]
-        package_ruby_ver: [ '3.0.6' ]
+        package_ruby_ver: [ '3.1.4' ]
         env:
           - CC: clang
             CXX: clang++
