+++ .github/workflows/macos.yml
@@ -91,6 +91,7 @@ jobs:
           brew install \
           bison flex binutils libffi double-conversion boost jemalloc fmt glog \
           gnu-sed bash pkg-config
+          echo "$(brew --prefix bison)/bin" >> $GITHUB_PATH
 
       - name: Install Ruby
         uses: ruby/setup-ruby@v1
@@ -107,7 +108,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ${{ github.workspace }}/${{ env.DEPS }}
-          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('**/common.env') }}-v${{ env.CACHE_VER }}
+          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('*.env') }}-v${{ env.CACHE_VER }}
 
       - name: tebako setup
         run: |
@@ -123,7 +124,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ macos-11.0 ]
-        package_ruby_ver: [ '3.0.6' ]
+        package_ruby_ver: [ '3.1.4', '3.2.2' ]
     steps:
       - name: Checkout tebako packaging environment
         uses: actions/checkout@v3
@@ -145,6 +146,7 @@ jobs:
           brew install \
           bison flex binutils libffi double-conversion boost jemalloc fmt glog \
           gnu-sed bash pkg-config
+          echo "$(brew --prefix bison)/bin" >> $GITHUB_PATH
 
       - name: Install Ruby
         uses: ruby/setup-ruby@v1
@@ -161,7 +163,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ${{ github.workspace }}/${{ env.DEPS }}
-          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('**/common.env') }}-v${{ env.CACHE_VER }}
+          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('*.env') }}-v${{ env.CACHE_VER }}
 
       - name: Cache check
         if: steps.cache.outputs.cache-hit != 'true'
@@ -202,7 +204,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ macos-11.0 ]
-        package_ruby_ver: [ '2.7.7', '3.0.6', '3.1.4' ]
+        package_ruby_ver: [ '3.1.4', '3.2.2' ]
     steps:
       - name: Checkout tebako packaging environment
         uses: actions/checkout@v3
@@ -224,6 +226,7 @@ jobs:
           brew install \
           bison flex binutils libffi double-conversion boost jemalloc fmt glog \
           gnu-sed bash pkg-config
+          echo "$(brew --prefix bison)/bin" >> $GITHUB_PATH
 
       - name: Install Ruby
         uses: ruby/setup-ruby@v1
@@ -240,7 +243,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ${{ github.workspace }}/${{ env.DEPS }}
-          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('**/common.env') }}-v${{ env.CACHE_VER }}
+          key: ${{ matrix.os }}-${{ env.CC }}-${{ hashFiles('*.env') }}-v${{ env.CACHE_VER }}
 
       - name: Cache check
         if: steps.cache.outputs.cache-hit != 'true'
