+++ README.adoc
@@ -217,7 +217,7 @@ invocation of press command.
 ----
 <install-folder>/exe/tebako setup     \
   [-p |--prefix=<tebako-root-folder>] \
-  [-R | --Ruby=<2.7.7|3.0.6|3.1.4>]
+  [-R | --Ruby=<2.7.7|3.0.6|3.1.4|3.2.2>]
 ----
 
 Where:
