+++ common.env
@@ -3,4 +3,4 @@ DEPS=deps
 INCBIN_TAG=348e36b
 DWARFS_WR_TAG=v0.3.4
 BUNDLER_VER=2.3.22
-RUBY_VER=3.0.6
+RUBY_VER=3.1.4
